requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-community
langchain-ollama
chromadb
tiktoken
unstructured
pdfminer-six
beautifulsoup4
requests
